bug: fix proxy cloning errs for basepath #232 (#233)
* bug: fix proxy cloning errs for basepath #232

* bug: fix proxy cloning errs proxy bundle #231

* format err handling #231

+++ cmd/apis/cloneapi.go
@@ -27,10 +27,9 @@ import (
 	"strings"
 
 	"internal/apiclient"
+	"internal/client/apis"
 
 	proxybundle "internal/bundlegen/proxybundle"
-
-	"internal/client/apis"
 
 	"github.com/spf13/cobra"
 )
@@ -67,11 +66,14 @@ var CloneCmd = &cobra.Command{
 		if proxyFolder != "" {
 			var tmpDir string
 
-			if tmpDir, err = copyDirectory(); err != nil {
-				return err
+			if proxyZip != "" {
+				tmpDir = proxyFolder
+			} else {
+				if tmpDir, err = copyDirectory(); err != nil {
+					return err
+				}
+				defer os.RemoveAll(tmpDir)
 			}
-
-			defer os.RemoveAll(tmpDir)
 
 			if err = renameProxy(tmpDir); err != nil {
 				return err
@@ -79,14 +81,14 @@ var CloneCmd = &cobra.Command{
 
 			proxyBundlePath := path.Join(tmpDir, name+".zip")
 
-			if err = proxybundle.GenerateArchiveBundle(proxyFolder, proxyBundlePath); err != nil {
+			if err = proxybundle.GenerateArchiveBundle(path.Join(tmpDir, "apiproxy"), proxyBundlePath); err != nil {
 				return err
 			}
 			if _, err = apis.CreateProxy(name, proxyBundlePath); err != nil {
 				return err
 			}
 
-			return os.Remove(proxyBundlePath)
+			return err
 		}
 
 		return err
@@ -159,7 +161,11 @@ func renameProxy(tmpDir string) (err error) {
 		return nil
 	})
 
-	return err
+	if err != nil {
+		return err
+	}
+
+	return setBasePath(path.Join(tmpDir, "apiproxy", "proxies", "default.xml"))
 }
 
 func setParam(filePath string, paramType string) (err error) {
@@ -179,16 +185,53 @@ func setParam(filePath string, paramType string) (err error) {
 
 	stringValue := string(byteValue)
 	replaceName := fmt.Sprintf("<APIProxy revision=\"1\" name=\"%s\">", name)
-	replaceBasePath := fmt.Sprintf("<BasePaths>%s</BasePaths>", basePath)
+	replaceBasePath := fmt.Sprintf("<Basepaths>%s</Basepaths>", basePath)
 
 	switch paramType {
 	case "proxy":
 		re := regexp.MustCompile(`\<APIProxy revision=\"\d+\" name=\"\S+\">`)
 		stringValue = re.ReplaceAllString(stringValue, replaceName)
 	case "basePath":
-		re := regexp.MustCompile(`\<Basepaths>\w+\<\\Basepaths>`)
+		re := regexp.MustCompile(`\<Basepaths>\/\w+\<\/Basepaths>`)
 		stringValue = re.ReplaceAllString(stringValue, replaceBasePath)
 	}
+
+	proxyFile, err = os.OpenFile(filePath, os.O_WRONLY|os.O_TRUNC, 0o644)
+	if err != nil {
+		return err
+	}
+
+	defer proxyFile.Close()
+
+	_, err = proxyFile.Write([]byte(stringValue))
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func setBasePath(filePath string) (err error) {
+	var proxyFile *os.File
+
+	replaceBasePath := fmt.Sprintf("<BasePath>%s</BasePath>", basePath)
+
+	proxyFile, err = os.OpenFile(filePath, os.O_RDONLY, 0o644)
+	if err != nil {
+		return err
+	}
+
+	byteValue, err := io.ReadAll(proxyFile)
+	if err != nil {
+		return err
+	}
+
+	proxyFile.Close()
+
+	stringValue := string(byteValue)
+
+	re := regexp.MustCompile(`\<BasePath>\/\w+\<\/BasePath>`)
+	stringValue = re.ReplaceAllString(stringValue, replaceBasePath)
 
 	proxyFile, err = os.OpenFile(filePath, os.O_WRONLY|os.O_TRUNC, 0o644)
 	if err != nil {
